# docs/swagger.yaml
...
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Testing Swagger APIs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: File Analyzer APIs
  version: "1.0"
paths:
  /getdata/{user_id}:
    get:
      description: Retrieves analyzed text data from the database for the given user
        ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Result data retrieved
        "404":
          description: Result not found
        "500":
          description: Database connection failed
      security:
      - BearerAuth: []
      summary: Get Analysis Result
      tags:
      - protected
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"